requirements.txt
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
Markdown==3.1
mysqlclient==1.4.2.post1
Pillow==5.4.1
Pygments==2.3.1
pytz==2018.9
